Move SigneeParty to Models folder
